generalize ProgramDiffPropertyTest to MergePropertyTest
